Initial Upload
Credit for the client goes to IKillKenny (IKK) years ago.
A lot of things work. A lot of things don't.
I'm still trying to understand how the battles relay data.

I spent so many hours playing Doc-Ent's games and without WotA, I'd not have made a few really great friends.

So I've worked on this off and on for a while and I'm just going to upload it in hopes that someone can help me out.

Do you have a more recent version of the client? Do you understand how the battles work? Take a stab at it or talk to me!